// src/generated/polycentric/v2/verifications_service.ts
export interface VerificationClaimBundle {
    /**
     * The VerificationClaim event.
     *
     * @generated from protobuf field: polycentric.v2.EventBundle claim = 1
     */
    claim?: EventBundle;
    /**
     * VerificationTarget events published by the claim's owner
     *
     * @generated from protobuf field: repeated polycentric.v2.EventBundle targets = 2
     */
    targets: EventBundle[];
    /**
     * VerificationVerify events.
     *
     * @generated from protobuf field: repeated polycentric.v2.EventBundle verifies = 3
     */
    verifies: EventBundle[];
}
// ...
class VerificationClaimBundle$Type extends MessageType<VerificationClaimBundle> {
    constructor() {
        super("polycentric.v2.VerificationClaimBundle", [
            { no: 1, name: "claim", kind: "message", T: () => EventBundle },
            { no: 2, name: "targets", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => EventBundle },
            { no: 3, name: "verifies", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => EventBundle }
        ]);
    }
    create(value?: PartialMessage<VerificationClaimBundle>): VerificationClaimBundle {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.targets = [];
        message.verifies = [];
        if (value !== undefined)
            reflectionMergePartial<VerificationClaimBundle>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: VerificationClaimBundle): VerificationClaimBundle {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* polycentric.v2.EventBundle claim */ 1:
                    message.claim = EventBundle.internalBinaryRead(reader, reader.uint32(), options, message.claim);
                    break;
                case /* repeated polycentric.v2.EventBundle targets */ 2:
                    message.targets.push(EventBundle.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* repeated polycentric.v2.EventBundle verifies */ 3:
                    message.verifies.push(EventBundle.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: VerificationClaimBundle, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* polycentric.v2.EventBundle claim = 1; */
        if (message.claim)
            EventBundle.internalBinaryWrite(message.claim, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* repeated polycentric.v2.EventBundle targets = 2; */
        for (let i = 0; i < message.targets.length; i++)
            EventBundle.internalBinaryWrite(message.targets[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* repeated polycentric.v2.EventBundle verifies = 3; */
        for (let i = 0; i < message.verifies.length; i++)
            EventBundle.internalBinaryWrite(message.verifies[i], writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message polycentric.v2.VerificationClaimBundle
 */
export const VerificationClaimBundle = new VerificationClaimBundle$Type();